adding "show members" to Spark
